Start preparing some normal mode movement keymaps
A "keymap" is just a function that gets a `{:buffer, :window}`
map, and returns an updated version of that map. The function
map also return just one of the keys, to indicate that the other
has unchanged. `keymaps/perform` builds this "context" map
and applies the changes to the provided cofx

Currently, a keymap is defined as a vec of keys. We might later simplify
this to a string, but vecs are pretty convenient