Scaffold out support for a completion popup menu
We have to patch ink because it [doesn't want to][1] support absolute
positioning out of the box.

We're probably *way* abusing ink at this point and should probably migrate
to [something else][2] but... I like flex. We'll see how far we can take
this before I give up, I guess.

[1]: https://github.com/vadimdemedes/ink/issues/182#issuecomment-483095617
[2]: https://github.com/vadimdemedes/ink/issues/182#issuecomment-2508669461