Reorganize a bit
The /build directory is sorta ignored, so moved our hooks into src/hooks
to avoid having to `git add -f` every time. `src/target` had a bit of a
smell, so moved the output to `target/ci-src` so all generated artifacts
live in `target` for consistency.

Last thing to fix is *either* simply run a program before compiling to
generate the ci entrypoint, or figure out the right way to update the
build-state so the generated cljs file gets compiled.