Set up CI (#4)
* WIP: Attempt to get tests to run

* Add a hard-coded test runner

Ideally we can generate the registry of test namespaces to import
instead...

* Generate saya.ci ns with a build hook

* Enable CI

* Reorganize a bit

The /build directory is sorta ignored, so moved our hooks into src/hooks
to avoid having to `git add -f` every time. `src/target` had a bit of a
smell, so moved the output to `target/ci-src` so all generated artifacts
live in `target` for consistency.

Last thing to fix is *either* simply run a program before compiling to
generate the ci entrypoint, or figure out the right way to update the
build-state so the generated cljs file gets compiled.

* Hackily add the ci file to the classpath + build-state

This works! *But* the file is output as `.js` (JUST that---only the file
extension) for... some reason?

* Clean up intentionally bad test

* Fix incorrect file name by adding :ns property

* Clean up unnecessary diff