Fix: input window now rendering inline properly
This might be a hack? But it works well enough. We need the :width to
fill the container for handling insert mode in non-connection buffers
without truncating, but we need it to collapse for rendering the input
window inline in connection buffers. There might be a better flex param
here, but this gets us unblocked....