Fix the off-by-one in derive-anchor-from-top-cursor
Pretty sure I should have already learned this, but we need to separate
the anchor-offset from the amount of "consumable lines." When not
wrapping, we have `:anchor-offset 0` and 1 consumable line. When wrapped
to 2, for example, we're at `:anchor-offset 1` and have 2 consumable
lines. This sounds like a simple `inc` but I'm pretty sure I tried that
at one point and it didn't work? Something something initial line
anchor-offset.

Scrolling is still not perfect---especially scrolling down towards the
bottom, and scrolling within a wrapped line---but I'm ready to just file
and move on for now.