Significantly improve scrolling + add tests (#12)
* Fix basic single-line (and window) scrolling

* Add non-wrapped cursor movement tests

* Add initial refactor to handle scrolling through wrapped lines

* Update window.subs-test for the new prop

* Clean up unused import

* Refactor adjust-cursor-to-scroll to handle virtual lines

* Make adjust-scroll-to-cursor always handle virtual lines

Our condition before was to restrictive!

* Perform adjust-scroll-to-cursor on window resize

* Add, update, and fix some tests

I should've confirmed how many virtual lines these got split into, that
would've helped a lot.

All tests are passing now *but* the behavior is still not quite right,
so we definitely need more tests...

* Improve the test helper with more context on errors

This helps us sanity check that our test is configured correctly

* *Finally* find a repro for scroll-to-top not working

* Fix the off-by-one in derive-anchor-from-top-cursor

Pretty sure I should have already learned this, but we need to separate
the anchor-offset from the amount of "consumable lines." When not
wrapping, we have `:anchor-offset 0` and 1 consumable line. When wrapped
to 2, for example, we're at `:anchor-offset 1` and have 2 consumable
lines. This sounds like a simple `inc` but I'm pretty sure I tried that
at one point and it didn't work? Something something initial line
anchor-offset.

Scrolling is still not perfect---especially scrolling down towards the
bottom, and scrolling within a wrapped line---but I'm ready to just file
and move on for now.

* Fix: use `visible-lines` to simplify adjust-cursor-to-scroll

* Preserve the "current" (virtual) column when clamping cursor