Scaffold basic support for scripting (#14)
* Scaffold basic `with-connection` support

Getting the connection ID is a bit janky, but... maybe okay? Need to add
the "are we already connected" safety check, and then fill out the rest

* Implement on-connection function, passing the connr

I've left the param name as `conn` for these functions in case we ever
decide we want to pass an object instead of just the raw connr

* Refactor connection management API + wire up script loading to CLI

We now have VERY rough CLI command parsing, supporting a
uri-or-script-path style of invocation.

* Fix the #send tagged literal + support multi-line

Didn't realize it was a macro! Also, you can now tag a collection
to send multiple lines in a row---super convenient for auto logins!

* Fix param ordering of perform-config invocation

Derp!

* Fix: gracefully handle *no* args again

Oops

* Track connection state, adjust setup-connection behavior accordingly

* Add a :reload command