Render the pending input buffer "dimmed" when not focused
This gives us a nice reminder that something is there!

Actually making this was kind of a pain. I'm not entirely sure why, but
if we have the "suffix-text" component in a separate k/Box from the
k/Text containing the cursor, 3 chars get truncated from the end of the
k/Box holding the buffer-line text. Very strange!

It's not *terribly* surprising given how much we're abusing ink here,
but... still annoying.