Refactor popup menu to use new upstream features (#30)
* Refactor the popup menu to use new Box :background-color support

Much cleaner! Would be lovely to be able to test this....

* Add a rough integration test for rendering the pum

Not sure why positioning the cursor at the bottom of the "screen"
renders all wacky in the test context... it seems totally fine in-app?

* Try stubbing out :stdin to see if that resolves CI issues

* Also stub stderr, just in case

* Fix some missing methods on our fake stdin

* Don't unmount until *after* extracting the output state

Seems to fix the *other* integration tests in CI, anyway.

* Use FORCE_COLOR=3 to ensure ansi is "supported" consistently in tests

This could be repro'd using `pnpm test | tail` to make chalk's default
terminal detection think there was no (or minimal) color support,
similar to whatever environment the github actions are run in.

This change ensures that tests relying on ansi coloring (like our
completion view test) all work locally, in editor/repl, and in CI.

* Bump min node version to 20, to match ink

* ci: Try throwing in a rerender?

* Add a separate test to prove verify absolute positioning works at all

Maybe the patch isn't applying in CI for some reason?

* Try bumping pnpm version to see if the older one skipped patch apply

Installing with frozen lockfile locally on this version seems to work
okay...