Fix input state persistence across new output (#31)
* Initialize connection input-window's cursor position from DB

* Prepare a unit test that repros the input state persistence issue

* Just persist input state on every change

Fixes #28

Basically the problem is that our cool "inline input" trick is actually
*reparenting* the input-window component into a new `[buffer-line]`
parent---this is not a type of movement that React supports.

Now ostensibly we could resolve this with Portals, but Portals don't
seem to be supported by ink AFAICT (normally they're part of the
ReactDOM API) so this is an okay 80/20.

I suppose what's odd is that we're already calling on-persist-value
on unmount, so you'd expect that the remount would "just work." My best
guess is that the remount happens *before* the delayed event-queue flush
from re-frame, resulting in the "new" input-window grabbing "old" state.

* Ignore "uninitialized kodachi" in unit tests