Strip null bytes from BufferLines
Fixes #34

Gotta be honest, I'm not really sure why this is happening, but it
*appears* that Procedural Realms sends a null byte (`0x00`) at the end
of its prompts. Stripping those out fixes the weird alignments
issues---and what are they doing there, anyway?

It looks like tt++ just [skips][1] them as well; it may be worth
tweaking things further to behave similarly (probably, upstream in
kodachi instead) but this unblocks us for now.

[1]: https://github.com/scandum/tintin/blob/cf8ac841e04f2628b3a5259e7b5605c3761b6735/src/telopt_server.c#L386