Throttle kodachi messages to apply in batch
This somewhat side-steps the perf issue with lines heavy in ansi colors,
but can still be helpful when receiving a large volume of lines (such as
when viewing a help document). One large kodachi dump I have went from
~4s to replay down to ~3s, while a smaller dump that took ~1.8s
dropped down to ~500ms.

The latter diff is somewhat artificial, since what it's really saving is
having to render a screenful of lines that are very heavy with ansi---if
you jump upward a page, all of those are forced to render at once,
resulting in a multi-second hang. Not ideal! But, I do think this is
still an improvement.

At higher throttle durations we can get even more gains, at the cost of
lower rendering. This seems like a fine balance for now; cranking it up
*could* help us focus our profiling efforts to optimize our handling of
heavy ansi lines...